+++ src/ProjectSystemTools/BuildLogging/Model/Backend/ILoggingDataSource.cs
@@ -2,7 +2,7 @@
 using System.Collections.Immutable;
 using Microsoft.VisualStudio.ProjectSystem.Tools.Providers;
 
-namespace Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.Backend
+namespace Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.BackEnd
 {
     public interface ILoggingDataSource
     {
