+++ src/ProjectSystemTools/BuildLogging/Model/Backend/ProjectLogger.cs
@@ -8,7 +8,7 @@ using Microsoft.Build.Framework;
 using Microsoft.Build.Logging;
 using Microsoft.VisualStudio.ProjectSystem.Tools.Providers;
 
-namespace Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.Backend
+namespace Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.BackEnd
 {
     internal sealed class ProjectLogger : LoggerBase
     {
@@ -20,7 +20,7 @@ namespace Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.Backend
         private readonly BinaryLogger _binaryLogger;
         private Build _build;
 
-        public ProjectLogger(BackendBuildTableDataSource dataSource, bool isDesignTime) :
+        public ProjectLogger(BackEndBuildTableDataSource dataSource, bool isDesignTime) :
             base(dataSource)
         {
             _logPath = Path.Combine(Path.GetTempPath(), $"{Guid.NewGuid()}.binlog");
