+++ src/ProjectSystemTools/BuildLogging/Model/Frontend/UIBuildSummary.cs
@@ -6,7 +6,7 @@ using Microsoft.VisualStudio.ProjectSystem.Tools.Providers;
 using Microsoft.VisualStudio.ProjectSystem.Tools.TableControl;
 using Microsoft.VisualStudio.Shell.TableManager;
 
-namespace Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.Frontend
+namespace Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.FrontEnd
 {
     /// <summary>
     /// Immutable Type
