+++ src/ProjectSystemTools/BuildLogging/Model/RpcContracts/IBuildLoggerService.cs
@@ -2,7 +2,7 @@
 
 using System;
 using System.Collections.Immutable;
-using Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.Backend;
+using Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.BackEnd;
 using Microsoft.VisualStudio.ProjectSystem.Tools.Providers;
 
 namespace Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.RpcContracts
