+++ src/ProjectSystemTools/BuildLogging/UI/BuildTableEntriesSnapshot.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System.Collections.Immutable;
-using Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.Frontend;
+using Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model.FrontEnd;
 using Microsoft.VisualStudio.Shell.TableControl;
 using Microsoft.VisualStudio.Shell.TableManager;
